Rearranged as a static library, because of errors (pointer arithmetics, typecast, ...) while compiling as a part of C++ project. CMakeLists.txt added.
