+++ libmdns.c
@@ -10,153 +10,9 @@
  * This library is put in the public domain; you can redistribute it and/or modify it without any restrictions.
  *
  */
+#include "libmdns.h"
 
-#pragma once
-
-#include <stdint.h>
-#include <stddef.h>
-#include <stdlib.h>
-#include <unistd.h>
-#include <string.h>
-
-#include <fcntl.h>
-#include <sys/socket.h>
-#include <netinet/in.h>
-
-#define MDNS_INVALID_POS ((size_t)-1)
-
-#define MDNS_STRING_CONST(s) (s), (sizeof((s))-1)
-#define MDNS_STRING_FORMAT(s) (int)((s).length), s.str
-
-enum mdns_record_type {
-	MDNS_RECORDTYPE_IGNORE = 0,
-	//Address
-	MDNS_RECORDTYPE_A = 1,
-	//Domain Name pointer
-	MDNS_RECORDTYPE_PTR = 12,
-	//Arbitrary text string
-	MDNS_RECORDTYPE_TXT = 16,
-	//IP6 Address [Thomson]
-	MDNS_RECORDTYPE_AAAA = 28,
-	//Server Selection [RFC2782]
-	MDNS_RECORDTYPE_SRV = 33
-};
-
-enum mdns_entry_type {
-	MDNS_ENTRYTYPE_ANSWER = 1,
-	MDNS_ENTRYTYPE_AUTHORITY = 2,
-	MDNS_ENTRYTYPE_ADDITIONAL = 3
-};
-
-enum mdns_class {
-	MDNS_CLASS_IN = 1
-};
-
-typedef enum mdns_record_type  mdns_record_type_t;
-typedef enum mdns_entry_type   mdns_entry_type_t;
-typedef enum mdns_class        mdns_class_t;
-
-typedef int (* mdns_record_callback_fn)(const struct sockaddr* from,
-                                        mdns_entry_type_t entry, uint16_t type,
-                                        uint16_t rclass, uint32_t ttl,
-                                        const void* data, size_t size, size_t offset, size_t length);
-
-typedef struct mdns_string_t       mdns_string_t;
-typedef struct mdns_string_pair_t  mdns_string_pair_t;
-typedef struct mdns_record_srv_t   mdns_record_srv_t;
-typedef struct mdns_record_txt_t   mdns_record_txt_t;
-
-struct mdns_string_t {
-	const char* str;
-	size_t length;
-};
-
-struct mdns_string_pair_t {
-	size_t  offset;
-	size_t  length;
-	int     ref;
-};
-
-struct mdns_record_srv_t {
-	uint16_t priority;
-	uint16_t weight;
-	uint16_t port;
-	mdns_string_t name;
-};
-
-struct mdns_record_txt_t {
-	mdns_string_t key;
-	mdns_string_t value;
-};
-
-static int
-mdns_socket_open_ipv4(void);
-
-static int
-mdns_socket_setup_ipv4(int sock);
-
-static int
-mdns_socket_open_ipv6(void);
-
-static int
-mdns_socket_setup_ipv6(int sock);
-
-static void
-mdns_socket_close(int sock);
-
-static int
-mdns_discovery_send(int sock);
-
-static size_t
-mdns_discovery_recv(int sock, void* buffer, size_t capacity,
-                    mdns_record_callback_fn callback);
-
-static int
-mdns_query_send(int sock, mdns_record_type_t type, const char* name, size_t length,
-                void* buffer, size_t capacity);
-
-static size_t
-mdns_query_recv(int sock, void* buffer, size_t capacity,
-                mdns_record_callback_fn callback);
-
-static mdns_string_t
-mdns_string_extract(const void* buffer, size_t size, size_t* offset,
-                    char* str, size_t capacity);
-
-static int
-mdns_string_skip(const void* buffer, size_t size, size_t* offset);
-
-static int
-mdns_string_equal(const void* buffer_lhs, size_t size_lhs, size_t* ofs_lhs,
-                  const void* buffer_rhs, size_t size_rhs, size_t* ofs_rhs);
-
-static void*
-mdns_string_make(void* data, size_t capacity, const char* name, size_t length);
-
-static mdns_string_t
-mdns_record_parse_ptr(const void* buffer, size_t size, size_t offset, size_t length,
-                      char* strbuffer, size_t capacity);
-
-static mdns_record_srv_t
-mdns_record_parse_srv(const void* buffer, size_t size, size_t offset, size_t length,
-                      char* strbuffer, size_t capacity);
-
-static struct sockaddr_in*
-mdns_record_parse_a(const void* buffer, size_t size, size_t offset, size_t length,
-                    struct sockaddr_in* addr);
-
-static struct sockaddr_in6*
-mdns_record_parse_aaaa(const void* buffer, size_t size, size_t offset, size_t length,
-                       struct sockaddr_in6* addr);
-
-static size_t
-mdns_record_parse_txt(const void* buffer, size_t size, size_t offset, size_t length,
-                      mdns_record_txt_t* records, size_t capacity);
-
-// Implementations
-
-static int
-mdns_socket_open_ipv4(void) {
+int mdns_socket_open_ipv4(void) {
 	int sock = (int)socket(AF_INET, SOCK_DGRAM, IPPROTO_UDP);
 	if (sock < 0)
 		return -1;
@@ -167,8 +23,7 @@ mdns_socket_open_ipv4(void) {
 	return sock;
 }
 
-static int
-mdns_socket_setup_ipv4(int sock) {
+int mdns_socket_setup_ipv4(int sock) {
 	struct sockaddr_in saddr;
 	memset(&saddr, 0, sizeof(saddr));
 	saddr.sin_family = AF_INET;
@@ -204,8 +59,7 @@ mdns_socket_setup_ipv4(int sock) {
 	return 0;
 }
 
-static int
-mdns_socket_open_ipv6(void) {
+int mdns_socket_open_ipv6(void) {
 	int sock = (int)socket(AF_INET6, SOCK_DGRAM, IPPROTO_UDP);
 	if (sock < 0)
 		return -1;
@@ -216,8 +70,7 @@ mdns_socket_open_ipv6(void) {
 	return sock;
 }
 
-static int
-mdns_socket_setup_ipv6(int sock) {
+int mdns_socket_setup_ipv6(int sock) {
 	struct sockaddr_in6 saddr;
 	memset(&saddr, 0, sizeof(saddr));
 	saddr.sin6_family = AF_INET6;
@@ -254,8 +107,7 @@ mdns_socket_setup_ipv6(int sock) {
 	return 0;
 }
 
-static void
-mdns_socket_close(int sock) {
+void mdns_socket_close(int sock) {
 #ifdef _WIN32
 	closesocket(sock);
 #else
@@ -263,13 +115,11 @@ mdns_socket_close(int sock) {
 #endif
 }
 
-static int
-mdns_is_string_ref(uint8_t val) {
+int mdns_is_string_ref(uint8_t val) {
 	return (0xC0 == (val & 0xC0));
 }
 
-static mdns_string_pair_t
-mdns_get_next_substring(const void* rawdata, size_t size, size_t offset) {
+mdns_string_pair_t mdns_get_next_substring(const void* rawdata, size_t size, size_t offset) {
 	const uint8_t* buffer = rawdata;
 	mdns_string_pair_t pair = {MDNS_INVALID_POS, 0, 0};
 	if (!buffer[offset]) {
@@ -297,8 +147,7 @@ mdns_get_next_substring(const void* rawdata, size_t size, size_t offset) {
 	return pair;
 }
 
-static int
-mdns_string_skip(const void* buffer, size_t size, size_t* offset) {
+int mdns_string_skip(const void* buffer, size_t size, size_t* offset) {
 	size_t cur = *offset;
 	mdns_string_pair_t substr;
 	do {
@@ -317,8 +166,7 @@ mdns_string_skip(const void* buffer, size_t size, size_t* offset) {
 	return 1;
 }
 
-static int
-mdns_string_equal(const void* buffer_lhs, size_t size_lhs, size_t* ofs_lhs,
+int mdns_string_equal(const void* buffer_lhs, size_t size_lhs, size_t* ofs_lhs,
                   const void* buffer_rhs, size_t size_rhs, size_t* ofs_rhs) {
 	size_t lhs_cur = *ofs_lhs;
 	size_t rhs_cur = *ofs_rhs;
@@ -356,8 +204,7 @@ mdns_string_equal(const void* buffer_lhs, size_t size_lhs, size_t* ofs_lhs,
 	return 1;
 }
 
-static mdns_string_t
-mdns_string_extract(const void* buffer, size_t size, size_t* offset,
+mdns_string_t mdns_string_extract(const void* buffer, size_t size, size_t* offset,
                     char* str, size_t capacity) {
 	size_t cur = *offset;
 	size_t end = MDNS_INVALID_POS;
@@ -393,8 +240,7 @@ mdns_string_extract(const void* buffer, size_t size, size_t* offset,
 	return result;
 }
 
-static size_t
-mdns_string_find(const char* str, size_t length, char c, size_t offset) {
+size_t mdns_string_find(const char* str, size_t length, char c, size_t offset) {
 	const void* found;
 	if (offset >= length)
 		return MDNS_INVALID_POS;
@@ -404,8 +250,7 @@ mdns_string_find(const char* str, size_t length, char c, size_t offset) {
 	return MDNS_INVALID_POS;
 }
 
-static void*
-mdns_string_make(void* data, size_t capacity, const char* name, size_t length) {
+void* mdns_string_make(void* data, size_t capacity, const char* name, size_t length) {
 	size_t pos = 0;
 	size_t last_pos = 0;
 	size_t remain = capacity;
@@ -441,8 +286,7 @@ mdns_string_make(void* data, size_t capacity, const char* name, size_t length) {
 	return dest;
 }
 
-static size_t
-mdns_records_parse(const struct sockaddr* from, const void* buffer, size_t size, size_t* offset,
+size_t mdns_records_parse(const struct sockaddr* from, const void* buffer, size_t size, size_t* offset,
                    mdns_entry_type_t type, size_t records, mdns_record_callback_fn callback) {
 	size_t parsed = 0;
 	int do_callback = 1;
@@ -491,8 +335,7 @@ static const uint8_t mdns_services_query[] = {
 	0x80, MDNS_CLASS_IN
 };
 
-static int
-mdns_discovery_send(int sock) {
+int mdns_discovery_send(int sock) {
 	struct sockaddr_in addr;
 	struct sockaddr_in6 addr6;
 	struct sockaddr* saddr = (struct sockaddr*)&addr;
@@ -530,8 +373,7 @@ mdns_discovery_send(int sock) {
 	return 0;
 }
 
-static size_t
-mdns_discovery_recv(int sock, void* buffer, size_t capacity,
+size_t mdns_discovery_recv(int sock, void* buffer, size_t capacity,
                     mdns_record_callback_fn callback) {
 	struct sockaddr_in6 addr;
 	struct sockaddr* saddr = (struct sockaddr*)&addr;
@@ -615,8 +457,7 @@ mdns_discovery_recv(int sock, void* buffer, size_t capacity,
 
 static uint16_t mdns_transaction_id = 0;
 
-static int
-mdns_query_send(int sock, mdns_record_type_t type, const char* name, size_t length,
+int mdns_query_send(int sock, mdns_record_type_t type, const char* name, size_t length,
                 void* buffer, size_t capacity) {
 	if (capacity < (17 + length))
 		return -1;
@@ -678,8 +519,7 @@ mdns_query_send(int sock, mdns_record_type_t type, const char* name, size_t leng
 	return 0;
 }
 
-static size_t
-mdns_query_recv(int sock, void* buffer, size_t capacity,
+size_t mdns_query_recv(int sock, void* buffer, size_t capacity,
                 mdns_record_callback_fn callback) {
 	struct sockaddr_in6 addr;
 	struct sockaddr* saddr = (struct sockaddr*)&addr;
@@ -732,8 +572,7 @@ mdns_query_recv(int sock, void* buffer, size_t capacity,
 	return records;
 }
 
-static mdns_string_t
-mdns_record_parse_ptr(const void* buffer, size_t size, size_t offset, size_t length,
+mdns_string_t mdns_record_parse_ptr(const void* buffer, size_t size, size_t offset, size_t length,
                       char* strbuffer, size_t capacity) {
 	//PTR record is just a string
 	if ((size >= offset + length) && (length >= 2))
@@ -742,8 +581,7 @@ mdns_record_parse_ptr(const void* buffer, size_t size, size_t offset, size_t len
 	return empty;
 }
 
-static mdns_record_srv_t
-mdns_record_parse_srv(const void* buffer, size_t size, size_t offset, size_t length,
+mdns_record_srv_t mdns_record_parse_srv(const void* buffer, size_t size, size_t offset, size_t length,
                       char* strbuffer, size_t capacity) {
 	mdns_record_srv_t srv;
 	memset(&srv, 0, sizeof(mdns_record_srv_t));
@@ -764,8 +602,7 @@ mdns_record_parse_srv(const void* buffer, size_t size, size_t offset, size_t len
 	return srv;
 }
 
-static struct sockaddr_in*
-mdns_record_parse_a(const void* buffer, size_t size, size_t offset, size_t length,
+struct sockaddr_in* mdns_record_parse_a(const void* buffer, size_t size, size_t offset, size_t length,
                     struct sockaddr_in* addr) {
 	memset(addr, 0, sizeof(struct sockaddr_in));
 	addr->sin_family = AF_INET;
@@ -777,8 +614,7 @@ mdns_record_parse_a(const void* buffer, size_t size, size_t offset, size_t lengt
 	return addr;
 }
 
-static struct sockaddr_in6*
-mdns_record_parse_aaaa(const void* buffer, size_t size, size_t offset, size_t length,
+struct sockaddr_in6* mdns_record_parse_aaaa(const void* buffer, size_t size, size_t offset, size_t length,
                        struct sockaddr_in6* addr) {
 	memset(addr, 0, sizeof(struct sockaddr_in6));
 	addr->sin6_family = AF_INET6;
@@ -790,8 +626,7 @@ mdns_record_parse_aaaa(const void* buffer, size_t size, size_t offset, size_t le
 	return addr;
 }
 
-static size_t
-mdns_record_parse_txt(const void* buffer, size_t size, size_t offset, size_t length,
+size_t mdns_record_parse_txt(const void* buffer, size_t size, size_t offset, size_t length,
                       mdns_record_txt_t* records, size_t capacity) {
 	size_t parsed = 0;
 	const char* strdata;
